Updated for 1.4. C# solution now uses .Net SDK format and PackageReferences for simplicity.
