+++ About/About.xml
@@ -5,6 +5,7 @@
   <supportedVersions>
     <li>1.2</li>
     <li>1.3</li>
+    <li>1.4</li>
   </supportedVersions>
   <packageId>JPT.keepitquiet</packageId>
   <modDependencies>
