--- Source/CompProperties_Soother.cs
@@ -1,14 +1,0 @@
-﻿using RimWorld;
-
-namespace KeepItQuiet
-{
-    public class CompProperties_Soother : CompProperties_Flickable
-    {
-        public CompProperties_Soother()
-        {
-            compClass = typeof(CompSoother);
-        }
-        public int power;
-        public int maxLevel;
-    }
-}
