+++ Source/KeepItQuiet.csproj.oldversioncscproj
