+++ Source/KeepItQuiet.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.5.33530.505
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "KeepItQuiet", "KeepItQuiet\KeepItQuiet.csproj", "{63B85605-72D9-4766-8774-A313BA818542}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{63B85605-72D9-4766-8774-A313BA818542}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{63B85605-72D9-4766-8774-A313BA818542}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{63B85605-72D9-4766-8774-A313BA818542}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{63B85605-72D9-4766-8774-A313BA818542}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
