+++ Source/KeepItQuiet/CompProperties_Soother.cs
@@ -1,0 +1,14 @@
+﻿using RimWorld;
+
+namespace KeepItQuiet;
+
+public class CompProperties_Soother : CompProperties_Flickable
+{
+    public int maxLevel;
+    public int power;
+
+    public CompProperties_Soother()
+    {
+        compClass = typeof(CompSoother);
+    }
+}
