+++ Source/KeepItQuiet/ModBaseKeepItQuiet.cs
@@ -1,0 +1,13 @@
+﻿using HugsLib;
+
+namespace KeepItQuiet;
+
+public class ModBaseKeepItQuiet : ModBase
+{
+    public ModBaseKeepItQuiet()
+    {
+        Settings.EntryName = "Keep It Quiet";
+    }
+
+    public override string ModIdentifier => "JPT_keepitquiet";
+}
