+++ Source/KeepItQuiet/NoisySoundDef.cs
@@ -1,0 +1,8 @@
+﻿using Verse;
+
+namespace KeepItQuiet;
+
+public class NoisySoundDef : SoundDef
+{
+    public int overrideNoiseVolume;
+}
