+++ Source/KeepItQuiet/QuietCategory.cs
@@ -1,0 +1,9 @@
+﻿namespace KeepItQuiet;
+
+public enum QuietCategory : byte
+{
+    Neutral,
+    Serene,
+    Disturbed,
+    DeeplyDisturbed
+}
