+++ Source/KeepItQuiet/QuietDefOf.cs
@@ -1,0 +1,14 @@
+﻿using RimWorld;
+using Verse;
+
+namespace KeepItQuiet;
+
+[DefOf]
+public static class QuietDefOf
+{
+    public static JobDef GotoQuietPlace;
+
+    public static TraitDef
+        NoiseSensitive,
+        NoiseTolerant;
+}
