--- Source/NoisySoundDef.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using Verse;
-
-namespace KeepItQuiet
-{
-    public class NoisySoundDef : SoundDef
-    {
-        public int overrideNoiseVolume;
-    }
-}
