--- Source/QuietCategory.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace KeepItQuiet
-{
-	public enum QuietCategory : byte
-	{
-		Neutral,
-		Serene,
-		Disturbed,
-		DeeplyDisturbed,
-	}
-}
