--- Source/QuietDefOf.cs
@@ -1,14 +1,0 @@
-﻿using RimWorld;
-using Verse;
-
-namespace KeepItQuiet
-{
-    [DefOf]
-    public static class QuietDefOf
-    {
-        public static JobDef GotoQuietPlace;
-        public static TraitDef
-            NoiseSensitive,
-            NoiseTolerant;
-    }
-}
